Don't allocate any memory during lookup.
The registered domain is always a suffix of the input domain, so we can
just walk backward through the input string until we find a match, then
return a pointer to the beginning of the match.  The matcher is now
shorter, nonrecursive, and more efficient.  This change also fixes the
test failure mentioned in rev e5231455fea59cc0b3b315ab16db027d7347c69f.

*API CHANGE:* Caller of getRegisteredDomain(Drop) now gets back a pointer
into the middle of the string it passed in, *not* a new malloced string.

This also removes the only use of strtok_r, which was the only non-C99
library routine used.